+++ src/API/Send/Classes/MailSenderNot4Factory.php
@@ -1,9 +1,9 @@
 <?php
 
-use SoerBot\API\Common;
-use SoerBot\API\Tools;
+use \API\Common;
+use \API\Tools;
 
-namespace SoerBot\API\Send;
+namespace \API\Send;
 /**
  * All classes are inherits from this not will register on factory
  */
@@ -19,4 +19,6 @@ class MailSenderNot4Factory extends AMailSender
     {
         return new Mail();
     }
+
+
 }
